properly enforce user permission
+ Properly enforce user permissions for uploading files, moving files
(file_edit), deleting folders (file_del), adding folders (file_add) on
the controller/script level. Not only on the template/view level.

+++ managefile.php
@@ -44,6 +44,13 @@ if (!chkproject($userid, $id)) {
     die();
 }
 if ($action == "upload") {
+    if (!$userpermissions["files"]["add"]) {
+        $errtxt = $langfile["nopermission"];
+        $noperm = $langfile["accessdenied"];
+        $template->assign("errortext", "<h2>$errtxt</h2><br>$noperm");
+        $template->display("error.tpl");
+        die();
+    }
     $num = $_POST['numfiles'];
 
     if ($upfolder) {
@@ -93,6 +100,13 @@ if ($action == "upload") {
     $loc = $url .= "managefile.php?action=showproject&id=$id&mode=added";
     header("Location: $loc");
 } elseif ($action == "uploadAsync") {
+    if (!$userpermissions["files"]["add"]) {
+        $errtxt = $langfile["nopermission"];
+        $noperm = $langfile["accessdenied"];
+        $template->assign("errortext", "<h2>$errtxt</h2><br>$noperm");
+        $template->display("error.tpl");
+        die();
+    }
     if ($upfolder) {
         $thefolder = $myfile->getFolder($upfolder);
         $thefolder = $thefolder["name"];
@@ -256,6 +270,13 @@ if ($action == "upload") {
     $template->assign("postmax", $POST_MAX_SIZE);
     $template->display("projectfiles.tpl");
 } elseif ($action == "addfolder") {
+    if (!$userpermissions["files"]["add"]) {
+        $errtxt = $langfile["nopermission"];
+        $noperm = $langfile["accessdenied"];
+        $template->assign("errortext", "<h2>$errtxt</h2><br>$noperm");
+        $template->display("error.tpl");
+        die();
+    }
     $name = getArrayVal($_POST, "foldertitle");
     $desc = getArrayVal($_POST, "folderdesc");
     $parent = getArrayVal($_POST, "folderparent");
@@ -268,6 +289,13 @@ if ($action == "upload") {
         header("Location: $loc");
     }
 } elseif ($action == "delfolder") {
+    if (!$userpermissions["files"]["del"]) {
+        $errtxt = $langfile["nopermission"];
+        $noperm = $langfile["accessdenied"];
+        $template->assign("errortext", "<h2>$errtxt</h2><br>$noperm");
+        $template->display("error.tpl");
+        die();
+    }
     $ajaxreq = $_GET["ajax"];
     $folder = getArrayVal($_GET, "folder");
     if ($myfile->deleteFolder($folder, $id)) {
@@ -279,6 +307,13 @@ if ($action == "upload") {
         }
     }
 } elseif ($action == "movefile") {
+    if (!$userpermissions["files"]["edit"]) {
+        $errtxt = $langfile["nopermission"];
+        $noperm = $langfile["accessdenied"];
+        $template->assign("errortext", "<h2>$errtxt</h2><br>$noperm");
+        $template->display("error.tpl");
+        die();
+    }
     $file = $_GET["file"];
     $file = substr($file, 4, strlen($file)-4);
